Network.AWS.Lightsail.DetachStaticIP

Description

Detaches a static IP from the Amazon Lightsail instance to which it is attached.

Creating a Request

detachStaticIP

Arguments

:: Text
dsiStaticIPName

-> DetachStaticIP

Creates a value of DetachStaticIP with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsiStaticIPName - The name of the static IP to detach from the instance.

data DetachStaticIP

See: detachStaticIP smart constructor.

Request Lenses

dsiStaticIPName :: Lens' DetachStaticIP Text

The name of the static IP to detach from the instance.

Destructuring the Response

detachStaticIPResponse

Arguments

:: Int
dsirsResponseStatus

-> DetachStaticIPResponse

Creates a value of DetachStaticIPResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsirsOperations - An array of key-value pairs containing information about the results of your detach static IP request.

dsirsResponseStatus - -- | The response status code.

data DetachStaticIPResponse

See: detachStaticIPResponse smart constructor.

Response Lenses

dsirsOperations :: Lens' DetachStaticIPResponse [Operation]

An array of key-value pairs containing information about the results of your detach static IP request.

dsirsResponseStatus :: Lens' DetachStaticIPResponse Int

- | The response status code.