Network.AWS.Lightsail.DownloadDefaultKeyPair

Description

Downloads the default SSH key pair from the user's account.

Creating a Request

downloadDefaultKeyPair :: DownloadDefaultKeyPair

Creates a value of DownloadDefaultKeyPair with the minimum fields required to make a request.

data DownloadDefaultKeyPair

See: downloadDefaultKeyPair smart constructor.

Destructuring the Response

downloadDefaultKeyPairResponse

Arguments

:: Int
ddkprsResponseStatus

-> DownloadDefaultKeyPairResponse

Creates a value of DownloadDefaultKeyPairResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddkprsPublicKeyBase64 - A base64-encoded public key of the ssh-rsa type.

ddkprsPrivateKeyBase64 - A base64-encoded RSA private key.

ddkprsResponseStatus - -- | The response status code.

data DownloadDefaultKeyPairResponse

See: downloadDefaultKeyPairResponse smart constructor.

Response Lenses

ddkprsPublicKeyBase64 :: Lens' DownloadDefaultKeyPairResponse (Maybe Text)

A base64-encoded public key of the ssh-rsa type.

ddkprsPrivateKeyBase64 :: Lens' DownloadDefaultKeyPairResponse (Maybe Text)

A base64-encoded RSA private key.

ddkprsResponseStatus :: Lens' DownloadDefaultKeyPairResponse Int

- | The response status code.