Network.AWS.Lightsail.GetBlueprints

Description

Returns the list of available instance images, or blueprints . You can use a blueprint to create a new virtual private server already running a specific operating system, as well as a preinstalled app or development stack. The software each instance is running depends on the blueprint image you choose.

This operation returns paginated results.

Creating a Request

getBlueprints :: GetBlueprints

Creates a value of GetBlueprints with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbIncludeInactive - A Boolean value indicating whether to include inactive results in your request.

gbPageToken - A token used for advancing to the next page of results from your get blueprints request.

data GetBlueprints

See: getBlueprints smart constructor.

Request Lenses

gbIncludeInactive :: Lens' GetBlueprints (Maybe Bool)

A Boolean value indicating whether to include inactive results in your request.

gbPageToken :: Lens' GetBlueprints (Maybe Text)

A token used for advancing to the next page of results from your get blueprints request.

Destructuring the Response

getBlueprintsResponse

Arguments

:: Int
gbsrsResponseStatus

-> GetBlueprintsResponse

Creates a value of GetBlueprintsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbsrsBlueprints - An array of key-value pairs that contains information about the available blueprints.

gbsrsNextPageToken - A token used for advancing to the next page of results from your get blueprints request.

gbsrsResponseStatus - -- | The response status code.

data GetBlueprintsResponse

See: getBlueprintsResponse smart constructor.

Response Lenses

gbsrsBlueprints :: Lens' GetBlueprintsResponse [Blueprint]

An array of key-value pairs that contains information about the available blueprints.

gbsrsNextPageToken :: Lens' GetBlueprintsResponse (Maybe Text)

A token used for advancing to the next page of results from your get blueprints request.

gbsrsResponseStatus :: Lens' GetBlueprintsResponse Int

- | The response status code.