Network.AWS.Lightsail.GetDomains

Description

Returns a list of all domains in the user's account.

This operation returns paginated results.

Creating a Request

getDomains :: GetDomains

Creates a value of GetDomains with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdPageToken - A token used for advancing to the next page of results from your get domains request.

data GetDomains

See: getDomains smart constructor.

Request Lenses

gdPageToken :: Lens' GetDomains (Maybe Text)

A token used for advancing to the next page of results from your get domains request.

Destructuring the Response

getDomainsResponse

Arguments

:: Int
gdsrsResponseStatus

-> GetDomainsResponse

Creates a value of GetDomainsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsrsNextPageToken - A token used for advancing to the next page of results from your get active names request.

gdsrsDomains - An array of key-value pairs containing information about each of the domain entries in the user's account.

gdsrsResponseStatus - -- | The response status code.

data GetDomainsResponse

See: getDomainsResponse smart constructor.

Response Lenses

gdsrsNextPageToken :: Lens' GetDomainsResponse (Maybe Text)

A token used for advancing to the next page of results from your get active names request.

gdsrsDomains :: Lens' GetDomainsResponse [Domain]

An array of key-value pairs containing information about each of the domain entries in the user's account.

gdsrsResponseStatus :: Lens' GetDomainsResponse Int

- | The response status code.