Network.AWS.Lightsail.GetInstanceMetricData

Description

Returns the data points for the specified Amazon Lightsail instance metric, given an instance name.

Creating a Request

getInstanceMetricData

Arguments

:: Text
gimdInstanceName

-> InstanceMetricName
gimdMetricName

-> Natural
gimdPeriod

-> UTCTime
gimdStartTime

-> UTCTime
gimdEndTime

-> MetricUnit
gimdUnit

-> GetInstanceMetricData

Creates a value of GetInstanceMetricData with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gimdInstanceName - The name of the instance for which you want to get metrics data.

gimdMetricName - The metric name to get data about.

gimdPeriod - The time period for which you are requesting data.

gimdStartTime - The start time of the time period.

gimdEndTime - The end time of the time period.

gimdUnit - The unit. The list of valid values is below.

gimdStatistics - The instance statistics.

data GetInstanceMetricData

See: getInstanceMetricData smart constructor.

Request Lenses

gimdInstanceName :: Lens' GetInstanceMetricData Text

The name of the instance for which you want to get metrics data.

gimdMetricName :: Lens' GetInstanceMetricData InstanceMetricName

The metric name to get data about.

gimdPeriod :: Lens' GetInstanceMetricData Natural

The time period for which you are requesting data.

gimdStartTime :: Lens' GetInstanceMetricData UTCTime

The start time of the time period.

gimdEndTime :: Lens' GetInstanceMetricData UTCTime

The end time of the time period.

gimdUnit :: Lens' GetInstanceMetricData MetricUnit

The unit. The list of valid values is below.

gimdStatistics :: Lens' GetInstanceMetricData [MetricStatistic]

The instance statistics.

Destructuring the Response

getInstanceMetricDataResponse

Arguments

:: Int
gimdrsResponseStatus

-> GetInstanceMetricDataResponse

Creates a value of GetInstanceMetricDataResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gimdrsMetricName - The metric name to return data for.

gimdrsMetricData - An array of key-value pairs containing information about the results of your get instance metric data request.

gimdrsResponseStatus - -- | The response status code.

data GetInstanceMetricDataResponse

See: getInstanceMetricDataResponse smart constructor.

Response Lenses

gimdrsMetricName :: Lens' GetInstanceMetricDataResponse (Maybe InstanceMetricName)

The metric name to return data for.

gimdrsMetricData :: Lens' GetInstanceMetricDataResponse [MetricDatapoint]

An array of key-value pairs containing information about the results of your get instance metric data request.

gimdrsResponseStatus :: Lens' GetInstanceMetricDataResponse Int

- | The response status code.