Network.AWS.Lightsail.GetInstancePortStates

Description

Returns the port states for a specific virtual private server, or instance .

Creating a Request

getInstancePortStates

Arguments

:: Text
gipsInstanceName

-> GetInstancePortStates

Creates a value of GetInstancePortStates with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gipsInstanceName - The name of the instance.

data GetInstancePortStates

See: getInstancePortStates smart constructor.

Request Lenses

gipsInstanceName :: Lens' GetInstancePortStates Text

The name of the instance.

Destructuring the Response

getInstancePortStatesResponse

Arguments

:: Int
gipsrsResponseStatus

-> GetInstancePortStatesResponse

Creates a value of GetInstancePortStatesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gipsrsPortStates - Information about the port states resulting from your request.

gipsrsResponseStatus - -- | The response status code.

data GetInstancePortStatesResponse

See: getInstancePortStatesResponse smart constructor.

Response Lenses

gipsrsPortStates :: Lens' GetInstancePortStatesResponse [InstancePortState]

Information about the port states resulting from your request.

gipsrsResponseStatus :: Lens' GetInstancePortStatesResponse Int

- | The response status code.