Network.AWS.Lightsail.GetInstanceSnapshot

Description

Returns information about a specific instance snapshot.

Creating a Request

getInstanceSnapshot

Arguments

:: Text
gisInstanceSnapshotName

-> GetInstanceSnapshot

Creates a value of GetInstanceSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gisInstanceSnapshotName - The name of the snapshot for which you are requesting information.

data GetInstanceSnapshot

See: getInstanceSnapshot smart constructor.

Request Lenses

gisInstanceSnapshotName :: Lens' GetInstanceSnapshot Text

The name of the snapshot for which you are requesting information.

Destructuring the Response

getInstanceSnapshotResponse

Arguments

:: Int
gisrsResponseStatus

-> GetInstanceSnapshotResponse

Creates a value of GetInstanceSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gisrsInstanceSnapshot - An array of key-value pairs containing information about the results of your get instance snapshot request.

gisrsResponseStatus - -- | The response status code.

data GetInstanceSnapshotResponse

See: getInstanceSnapshotResponse smart constructor.

Response Lenses

gisrsInstanceSnapshot :: Lens' GetInstanceSnapshotResponse (Maybe InstanceSnapshot)

An array of key-value pairs containing information about the results of your get instance snapshot request.

gisrsResponseStatus :: Lens' GetInstanceSnapshotResponse Int

- | The response status code.