Network.AWS.Lightsail.GetInstanceSnapshots

Description

Returns all instance snapshots for the user's account.

This operation returns paginated results.

Creating a Request

getInstanceSnapshots :: GetInstanceSnapshots

Creates a value of GetInstanceSnapshots with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gisPageToken - A token used for advancing to the next page of results from your get instance snapshots request.

data GetInstanceSnapshots

See: getInstanceSnapshots smart constructor.

Request Lenses

gisPageToken :: Lens' GetInstanceSnapshots (Maybe Text)

A token used for advancing to the next page of results from your get instance snapshots request.

Destructuring the Response

getInstanceSnapshotsResponse

Arguments

:: Int
gissrsResponseStatus

-> GetInstanceSnapshotsResponse

Creates a value of GetInstanceSnapshotsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gissrsNextPageToken - A token used for advancing to the next page of results from your get instance snapshots request.

gissrsInstanceSnapshots - An array of key-value pairs containing information about the results of your get instance snapshots request.

gissrsResponseStatus - -- | The response status code.

data GetInstanceSnapshotsResponse

See: getInstanceSnapshotsResponse smart constructor.

Response Lenses

gissrsNextPageToken :: Lens' GetInstanceSnapshotsResponse (Maybe Text)

A token used for advancing to the next page of results from your get instance snapshots request.

gissrsInstanceSnapshots :: Lens' GetInstanceSnapshotsResponse [InstanceSnapshot]

An array of key-value pairs containing information about the results of your get instance snapshots request.

gissrsResponseStatus :: Lens' GetInstanceSnapshotsResponse Int

- | The response status code.