Network.AWS.Lightsail.GetInstances

Description

Returns information about all Amazon Lightsail virtual private servers, or instances .

This operation returns paginated results.

Creating a Request

getInstances :: GetInstances

Creates a value of GetInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giPageToken - A token used for advancing to the next page of results from your get instances request.

data GetInstances

See: getInstances smart constructor.

Request Lenses

giPageToken :: Lens' GetInstances (Maybe Text)

A token used for advancing to the next page of results from your get instances request.

Destructuring the Response

getInstancesResponse

Arguments

:: Int
grsResponseStatus

-> GetInstancesResponse

Creates a value of GetInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grsNextPageToken - A token used for advancing to the next page of results from your get instances request.

grsInstances - An array of key-value pairs containing information about your instances.

grsResponseStatus - -- | The response status code.

data GetInstancesResponse

See: getInstancesResponse smart constructor.

Response Lenses

grsNextPageToken :: Lens' GetInstancesResponse (Maybe Text)

A token used for advancing to the next page of results from your get instances request.

grsInstances :: Lens' GetInstancesResponse [Instance]

An array of key-value pairs containing information about your instances.

grsResponseStatus :: Lens' GetInstancesResponse Int

- | The response status code.