Network.AWS.Lightsail.GetKeyPair

Description

Returns information about a specific key pair.

Creating a Request

getKeyPair

Arguments

:: Text
gkpKeyPairName

-> GetKeyPair

Creates a value of GetKeyPair with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gkpKeyPairName - The name of the key pair for which you are requesting information.

data GetKeyPair

See: getKeyPair smart constructor.

Request Lenses

gkpKeyPairName :: Lens' GetKeyPair Text

The name of the key pair for which you are requesting information.

Destructuring the Response

getKeyPairResponse

Arguments

:: Int
gkprsResponseStatus

-> GetKeyPairResponse

Creates a value of GetKeyPairResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gkprsKeyPair - An array of key-value pairs containing information about the key pair.

gkprsResponseStatus - -- | The response status code.

data GetKeyPairResponse

See: getKeyPairResponse smart constructor.

Response Lenses

gkprsKeyPair :: Lens' GetKeyPairResponse (Maybe KeyPair)

An array of key-value pairs containing information about the key pair.

gkprsResponseStatus :: Lens' GetKeyPairResponse Int

- | The response status code.