Network.AWS.Lightsail.GetKeyPairs

Description

Returns information about all key pairs in the user's account.

This operation returns paginated results.

Creating a Request

getKeyPairs :: GetKeyPairs

Creates a value of GetKeyPairs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gkpPageToken - A token used for advancing to the next page of results from your get key pairs request.

data GetKeyPairs

See: getKeyPairs smart constructor.

Request Lenses

gkpPageToken :: Lens' GetKeyPairs (Maybe Text)

A token used for advancing to the next page of results from your get key pairs request.

Destructuring the Response

getKeyPairsResponse

Arguments

:: Int
gkpsrsResponseStatus

-> GetKeyPairsResponse

Creates a value of GetKeyPairsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gkpsrsNextPageToken - A token used for advancing to the next page of results from your get key pairs request.

gkpsrsKeyPairs - An array of key-value pairs containing information about the key pairs.

gkpsrsResponseStatus - -- | The response status code.

data GetKeyPairsResponse

See: getKeyPairsResponse smart constructor.

Response Lenses

gkpsrsNextPageToken :: Lens' GetKeyPairsResponse (Maybe Text)

A token used for advancing to the next page of results from your get key pairs request.

gkpsrsKeyPairs :: Lens' GetKeyPairsResponse [KeyPair]

An array of key-value pairs containing information about the key pairs.

gkpsrsResponseStatus :: Lens' GetKeyPairsResponse Int

- | The response status code.