Network.AWS.Lightsail.GetOperations

Description

Returns information about all operations.

Results are returned from oldest to newest, up to a maximum of 200. Results can be paged by making each subsequent call to GetOperations use the maximum (last) statusChangedAt value from the previous request.

This operation returns paginated results.

Creating a Request

getOperations :: GetOperations

Creates a value of GetOperations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

goPageToken - A token used for advancing to the next page of results from your get operations request.

data GetOperations

See: getOperations smart constructor.

Request Lenses

goPageToken :: Lens' GetOperations (Maybe Text)

A token used for advancing to the next page of results from your get operations request.

Destructuring the Response

getOperationsResponse

Arguments

:: Int
gosrsResponseStatus

-> GetOperationsResponse

Creates a value of GetOperationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gosrsNextPageToken - A token used for advancing to the next page of results from your get operations request.

gosrsOperations - An array of key-value pairs containing information about the results of your get operations request.

gosrsResponseStatus - -- | The response status code.

data GetOperationsResponse

See: getOperationsResponse smart constructor.

Response Lenses

gosrsNextPageToken :: Lens' GetOperationsResponse (Maybe Text)

A token used for advancing to the next page of results from your get operations request.

gosrsOperations :: Lens' GetOperationsResponse [Operation]

An array of key-value pairs containing information about the results of your get operations request.

gosrsResponseStatus :: Lens' GetOperationsResponse Int

- | The response status code.