Network.AWS.Lightsail.GetOperationsForResource

Description

Gets operations for a specific resource (e.g., an instance or a static IP).

Creating a Request

getOperationsForResource

Arguments

:: Text
gofrResourceName

-> GetOperationsForResource

Creates a value of GetOperationsForResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gofrPageToken - A token used for advancing to the next page of results from your get operations for resource request.

gofrResourceName - The name of the resource for which you are requesting information.

data GetOperationsForResource

See: getOperationsForResource smart constructor.

Request Lenses

gofrPageToken :: Lens' GetOperationsForResource (Maybe Text)

A token used for advancing to the next page of results from your get operations for resource request.

gofrResourceName :: Lens' GetOperationsForResource Text

The name of the resource for which you are requesting information.

Destructuring the Response

getOperationsForResourceResponse

Arguments

:: Int
gofrrsResponseStatus

-> GetOperationsForResourceResponse

Creates a value of GetOperationsForResourceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gofrrsNextPageCount - (Deprecated) Returns the number of pages of results that remain.

gofrrsNextPageToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

gofrrsOperations - An array of key-value pairs containing information about the results of your get operations for resource request.

gofrrsResponseStatus - -- | The response status code.

data GetOperationsForResourceResponse

See: getOperationsForResourceResponse smart constructor.

Response Lenses

gofrrsNextPageCount :: Lens' GetOperationsForResourceResponse (Maybe Text)

(Deprecated) Returns the number of pages of results that remain.

gofrrsNextPageToken :: Lens' GetOperationsForResourceResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

gofrrsOperations :: Lens' GetOperationsForResourceResponse [Operation]

An array of key-value pairs containing information about the results of your get operations for resource request.

gofrrsResponseStatus :: Lens' GetOperationsForResourceResponse Int

- | The response status code.