Network.AWS.Lightsail.GetStaticIP

Description

Returns information about a specific static IP.

Creating a Request

getStaticIP

Arguments

:: Text
gsiStaticIPName

-> GetStaticIP

Creates a value of GetStaticIP with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsiStaticIPName - The name of the static IP in Lightsail.

data GetStaticIP

See: getStaticIP smart constructor.

Request Lenses

gsiStaticIPName :: Lens' GetStaticIP Text

The name of the static IP in Lightsail.

Destructuring the Response

getStaticIPResponse

Arguments

:: Int
gsirsResponseStatus

-> GetStaticIPResponse

Creates a value of GetStaticIPResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsirsStaticIP - An array of key-value pairs containing information about the requested static IP.

gsirsResponseStatus - -- | The response status code.

data GetStaticIPResponse

See: getStaticIPResponse smart constructor.

Response Lenses

gsirsStaticIP :: Lens' GetStaticIPResponse (Maybe StaticIP)

An array of key-value pairs containing information about the requested static IP.

gsirsResponseStatus :: Lens' GetStaticIPResponse Int

- | The response status code.