Network.AWS.Lightsail.ImportKeyPair

Description

Imports a public SSH key from a specific key pair.

Creating a Request

importKeyPair

Arguments

:: Text
ikpKeyPairName

-> Text
ikpPublicKeyBase64

-> ImportKeyPair

Creates a value of ImportKeyPair with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ikpKeyPairName - The name of the key pair for which you want to import the public key.

ikpPublicKeyBase64 - A base64-encoded public key of the ssh-rsa type.

data ImportKeyPair

See: importKeyPair smart constructor.

Request Lenses

ikpKeyPairName :: Lens' ImportKeyPair Text

The name of the key pair for which you want to import the public key.

ikpPublicKeyBase64 :: Lens' ImportKeyPair Text

A base64-encoded public key of the ssh-rsa type.

Destructuring the Response

importKeyPairResponse

Arguments

:: Int
ikprsResponseStatus

-> ImportKeyPairResponse

Creates a value of ImportKeyPairResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ikprsOperation - An array of key-value pairs containing information about the request operation.

ikprsResponseStatus - -- | The response status code.

data ImportKeyPairResponse

See: importKeyPairResponse smart constructor.

Response Lenses

ikprsOperation :: Lens' ImportKeyPairResponse (Maybe Operation)

An array of key-value pairs containing information about the request operation.

ikprsResponseStatus :: Lens' ImportKeyPairResponse Int

- | The response status code.