Network.AWS.Lightsail.Types

Description

Service Configuration

lightsail :: Service

API version 2016-11-28 of the Amazon Lightsail SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Lightsail throws this exception when the user cannot be authenticated or uses invalid credentials to access a resource.

_AccountSetupInProgressException :: AsError a => Getting (First ServiceError) a ServiceError

Lightsail throws this exception when an account is still in the setup in progress state.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Lightsail throws this exception when it cannot find a resource.

_OperationFailureException :: AsError a => Getting (First ServiceError) a ServiceError

Lightsail throws this exception when an operation fails to execute.

_ServiceException :: AsError a => Getting (First ServiceError) a ServiceError

A general service exception.

_UnauthenticatedException :: AsError a => Getting (First ServiceError) a ServiceError

Lightsail throws this exception when the user has not been authenticated.

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

Lightsail throws this exception when user input does not conform to the validation rules of an input field.

AccessDirection

data AccessDirection

Constructors

Inbound

Outbound

BlueprintType

data BlueprintType

Constructors

App

OS

InstanceAccessProtocol

data InstanceAccessProtocol

Constructors

Rdp

SSH

InstanceMetricName

data InstanceMetricName

Constructors

CPUUtilization

NetworkIn

NetworkOut

StatusCheckFailed

StatusCheckFailedInstance

StatusCheckFailedSystem

InstanceSnapshotState

data InstanceSnapshotState

Constructors

Available

Error'

Pending

MetricStatistic

data MetricStatistic

Constructors

Average

Maximum

Minimum

SampleCount

Sum

MetricUnit

data MetricUnit

Constructors

Bits

BitsSecond

Bytes

BytesSecond

Count

CountSecond

Gigabits

GigabitsSecond

Gigabytes

GigabytesSecond

Kilobits

KilobitsSecond

Kilobytes

KilobytesSecond

Megabits

MegabitsSecond

Megabytes

MegabytesSecond

Microseconds

Milliseconds

None

Percent

Seconds

Terabits

TerabitsSecond

Terabytes

TerabytesSecond

NetworkProtocol

data NetworkProtocol

Constructors

All

TCP

Udp

OperationStatus

data OperationStatus

Constructors

Completed

Failed

NotStarted

Started

OperationType

data OperationType

Constructors

AllocateStaticIP

AttachStaticIP

CloseInstancePublicPorts

CreateDomain

CreateInstance

CreateInstanceSnapshot

CreateInstancesFromSnapshot

DeleteDomain

DeleteDomainEntry

DeleteInstance

DeleteInstanceSnapshot

DetachStaticIP

OpenInstancePublicPorts

PutInstancePublicPorts

RebootInstance

ReleaseStaticIP

StartInstance

StopInstance

UpdateDomainEntry

PortAccessType

data PortAccessType

Constructors

Private

Public

PortState

data PortState

Constructors

Closed

Open

RegionName

data RegionName

Constructors

ApNortheast1

ApNortheast2

ApSouth1

ApSoutheast1

ApSoutheast2

EuCentral1

EuWest1

UsEast1

UsEast2

UsWest1

UsWest2

ResourceType

data ResourceType

Constructors

Domain

Instance

InstanceSnapshot

KeyPair

PeeredVPC

StaticIP

AvailabilityZone

data AvailabilityZone

Describes an Availability Zone.

See: availabilityZone smart constructor.

availabilityZone :: AvailabilityZone

Creates a value of AvailabilityZone with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

azState - The state of the Availability Zone.

azZoneName - The name of the Availability Zone. The format is us-east-1a (case-sensitive).

azState :: Lens' AvailabilityZone (Maybe Text)

The state of the Availability Zone.

azZoneName :: Lens' AvailabilityZone (Maybe Text)

The name of the Availability Zone. The format is us-east-1a (case-sensitive).

Blueprint

data Blueprint

Describes a blueprint (a virtual private server image).

See: blueprint smart constructor.

blueprint :: Blueprint

Creates a value of Blueprint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bVersionCode - The version code.

bGroup - The group name of the blueprint (e.g., amazon-linux ).

bMinPower - The minimum machine size required to run this blueprint. 0 indicates that the blueprint runs on all instances.

bProductURL - The product URL to learn more about the image or blueprint.

bLicenseURL - The end-user license agreement URL for the image or blueprint.

bName - The friendly name of the blueprint (e.g., Amazon Linux ).

bVersion - The version number of the operating system, application, or stack (e.g., 2016.03.0 ).

bBlueprintId - The ID for the virtual private server image (e.g., app_wordpress_4_4 or app_lamp_7_0 ).

bType - The type of the blueprint (e.g., os or app ).

bIsActive - A Boolean value indicating whether the blueprint is active. When you update your blueprints, you will inactivate old blueprints and keep the most recent versions active.

bDescription - The description of the blueprint.

bVersionCode :: Lens' Blueprint (Maybe Text)

The version code.

bGroup :: Lens' Blueprint (Maybe Text)

The group name of the blueprint (e.g., amazon-linux ).

bMinPower :: Lens' Blueprint (Maybe Int)

The minimum machine size required to run this blueprint. 0 indicates that the blueprint runs on all instances.

bProductURL :: Lens' Blueprint (Maybe Text)

The product URL to learn more about the image or blueprint.

bLicenseURL :: Lens' Blueprint (Maybe Text)

The end-user license agreement URL for the image or blueprint.

bName :: Lens' Blueprint (Maybe Text)

The friendly name of the blueprint (e.g., Amazon Linux ).

bVersion :: Lens' Blueprint (Maybe Text)

The version number of the operating system, application, or stack (e.g., 2016.03.0 ).

bBlueprintId :: Lens' Blueprint (Maybe Text)

The ID for the virtual private server image (e.g., app_wordpress_4_4 or app_lamp_7_0 ).

bType :: Lens' Blueprint (Maybe BlueprintType)

The type of the blueprint (e.g., os or app ).

bIsActive :: Lens' Blueprint (Maybe Bool)

A Boolean value indicating whether the blueprint is active. When you update your blueprints, you will inactivate old blueprints and keep the most recent versions active.

bDescription :: Lens' Blueprint (Maybe Text)

The description of the blueprint.

Bundle

data Bundle

Describes a bundle, which is a set of specs describing your virtual private server (or instance ).

See: bundle smart constructor.

bundle :: Bundle

Creates a value of Bundle with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bunCpuCount - The number of vCPUs included in the bundle (e.g., 2 ).

bunTransferPerMonthInGb - The data transfer rate per month in GB (e.g., 2000 ).

bunBundleId - The bundle ID (e.g., micro_1_0 ).

bunInstanceType - The Amazon EC2 instance type (e.g., t2.micro ).

bunName - A friendly name for the bundle (e.g., Micro ).

bunPower - The power of the bundle (e.g., 500 ).

bunDiskSizeInGb - The size of the SSD (e.g., 30 ).

bunPrice - The price in US dollars (e.g., 5.0 ).

bunIsActive - A Boolean value indicating whether the bundle is active.

bunRamSizeInGb - The amount of RAM in GB (e.g., 2.0 ).

bunCpuCount :: Lens' Bundle (Maybe Int)

The number of vCPUs included in the bundle (e.g., 2 ).

bunTransferPerMonthInGb :: Lens' Bundle (Maybe Int)

The data transfer rate per month in GB (e.g., 2000 ).

bunBundleId :: Lens' Bundle (Maybe Text)

The bundle ID (e.g., micro_1_0 ).

bunInstanceType :: Lens' Bundle (Maybe Text)

The Amazon EC2 instance type (e.g., t2.micro ).

bunName :: Lens' Bundle (Maybe Text)

A friendly name for the bundle (e.g., Micro ).

bunPower :: Lens' Bundle (Maybe Int)

The power of the bundle (e.g., 500 ).

bunDiskSizeInGb :: Lens' Bundle (Maybe Int)

The size of the SSD (e.g., 30 ).

bunPrice :: Lens' Bundle (Maybe Double)

The price in US dollars (e.g., 5.0 ).

bunIsActive :: Lens' Bundle (Maybe Bool)

A Boolean value indicating whether the bundle is active.

bunRamSizeInGb :: Lens' Bundle (Maybe Double)

The amount of RAM in GB (e.g., 2.0 ).

Disk

data Disk

Describes the hard disk (an SSD).

See: disk smart constructor.

disk :: Disk

Creates a value of Disk with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dResourceType - The resource type of the disk.

dArn - The Amazon Resource Name (ARN) of the disk.

dPath - The disk path.

dCreatedAt - The date when the disk was created.

dLocation - The region and Availability Zone where the disk is located.

dIops - The input/output operations per second (IOPS) of the disk.

dIsAttached - A Boolean value indicating whether the disk is attached.

dAttachmentState - The attachment state of the disk.

dName - The name of the disk.

dSizeInGb - The size of the disk in GB.

dSupportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

dIsSystemDisk - A Boolean value indicating whether this disk is a system disk (has an operating system loaded on it).

dAttachedTo - The resources to which the disk is attached.

dGbInUse - The number of GB in use by the disk.

dResourceType :: Lens' Disk (Maybe ResourceType)

The resource type of the disk.

dArn :: Lens' Disk (Maybe Text)

The Amazon Resource Name (ARN) of the disk.

dPath :: Lens' Disk (Maybe Text)

The disk path.

dCreatedAt :: Lens' Disk (Maybe UTCTime)

The date when the disk was created.

dLocation :: Lens' Disk (Maybe ResourceLocation)

The region and Availability Zone where the disk is located.

dIops :: Lens' Disk (Maybe Int)

The input/output operations per second (IOPS) of the disk.

dIsAttached :: Lens' Disk (Maybe Bool)

A Boolean value indicating whether the disk is attached.

dAttachmentState :: Lens' Disk (Maybe Text)

The attachment state of the disk.

dName :: Lens' Disk (Maybe Text)

The name of the disk.

dSizeInGb :: Lens' Disk (Maybe Int)

The size of the disk in GB.

dSupportCode :: Lens' Disk (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

dIsSystemDisk :: Lens' Disk (Maybe Bool)

A Boolean value indicating whether this disk is a system disk (has an operating system loaded on it).

dAttachedTo :: Lens' Disk (Maybe Text)

The resources to which the disk is attached.

dGbInUse :: Lens' Disk (Maybe Int)

The number of GB in use by the disk.

Domain

data Domain

Describes a domain where you are storing recordsets in Lightsail.

See: domain smart constructor.

domain :: Domain

Creates a value of Domain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

domResourceType - The resource type.

domDomainEntries - An array of key-value pairs containing information about the domain entries.

domArn - The Amazon Resource Name (ARN) of the domain recordset (e.g., arn:aws:lightsail:global:123456789101:Domain/824cede0-abc7-4f84-8dbc-12345EXAMPLE ).

domCreatedAt - The date when the domain recordset was created.

domLocation - The AWS Region and Availability Zones where the domain recordset was created.

domName - The name of the domain.

domSupportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

domResourceType :: Lens' Domain (Maybe ResourceType)

The resource type.

domDomainEntries :: Lens' Domain [DomainEntry]

An array of key-value pairs containing information about the domain entries.

domArn :: Lens' Domain (Maybe Text)

The Amazon Resource Name (ARN) of the domain recordset (e.g., arn:aws:lightsail:global:123456789101:Domain/824cede0-abc7-4f84-8dbc-12345EXAMPLE ).

domCreatedAt :: Lens' Domain (Maybe UTCTime)

The date when the domain recordset was created.

domLocation :: Lens' Domain (Maybe ResourceLocation)

The AWS Region and Availability Zones where the domain recordset was created.

domName :: Lens' Domain (Maybe Text)

The name of the domain.

domSupportCode :: Lens' Domain (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

DomainEntry

data DomainEntry

Describes a domain recordset entry.

See: domainEntry smart constructor.

domainEntry :: DomainEntry

Creates a value of DomainEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deName - The name of the domain.

deId - The ID of the domain recordset entry.

deOptions - The options for the domain entry.

deType - The type of domain entry (e.g., SOA or NS ).

deTarget - The target AWS name server (e.g., ns-111.awsdns-22.com. ).

deName :: Lens' DomainEntry (Maybe Text)

The name of the domain.

deId :: Lens' DomainEntry (Maybe Text)

The ID of the domain recordset entry.

deOptions :: Lens' DomainEntry (HashMap Text Text)

The options for the domain entry.

deType :: Lens' DomainEntry (Maybe Text)

The type of domain entry (e.g., SOA or NS ).

deTarget :: Lens' DomainEntry (Maybe Text)

The target AWS name server (e.g., ns-111.awsdns-22.com. ).

Instance

data Instance

Describes an instance (a virtual private server).

See: instance' smart constructor.

instance' :: Instance

Creates a value of Instance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iState - The status code and the state (e.g., running ) for the instance.

iIpv6Address - The IPv6 address of the instance.

iResourceType - The type of resource (usually Instance ).

iArn - The Amazon Resource Name (ARN) of the instance (e.g., arn:aws:lightsail:us-east-1:123456789101:Instance/244ad76f-8aad-4741-809f-12345EXAMPLE ).

iCreatedAt - The timestamp when the instance was created (e.g., 1479734909.17 ).

iLocation - The region name and availability zone where the instance is located.

iSshKeyName - The name of the SSH key being used to connect to the instance (e.g., LightsailDefaultKeyPair ).

iUsername - The user name for connecting to the instance (e.g., ec2-user ).

iNetworking - Information about the public ports and monthly data transfer rates for the instance.

iBundleId - The bundle for the instance (e.g., micro_1_0 ).

iName - The name the user gave the instance (e.g., Amazon_Linux-1GB-Virginia-1 ).

iSupportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

iBlueprintId - The blueprint ID (e.g., os_amlinux_2016_03 ).

iPrivateIPAddress - The private IP address of the instance.

iBlueprintName - The friendly name of the blueprint (e.g., Amazon Linux ).

iIsStaticIP - A Boolean value indicating whether this instance has a static IP assigned to it.

iPublicIPAddress - The public IP address of the instance.

iHardware - The size of the vCPU and the amount of RAM for the instance.

iState :: Lens' Instance (Maybe InstanceState)

The status code and the state (e.g., running ) for the instance.

iIpv6Address :: Lens' Instance (Maybe Text)

The IPv6 address of the instance.

iResourceType :: Lens' Instance (Maybe ResourceType)

The type of resource (usually Instance ).

iArn :: Lens' Instance (Maybe Text)

The Amazon Resource Name (ARN) of the instance (e.g., arn:aws:lightsail:us-east-1:123456789101:Instance/244ad76f-8aad-4741-809f-12345EXAMPLE ).

iCreatedAt :: Lens' Instance (Maybe UTCTime)

The timestamp when the instance was created (e.g., 1479734909.17 ).

iLocation :: Lens' Instance (Maybe ResourceLocation)

The region name and availability zone where the instance is located.

iSshKeyName :: Lens' Instance (Maybe Text)

The name of the SSH key being used to connect to the instance (e.g., LightsailDefaultKeyPair ).

iUsername :: Lens' Instance (Maybe Text)

The user name for connecting to the instance (e.g., ec2-user ).

iNetworking :: Lens' Instance (Maybe InstanceNetworking)

Information about the public ports and monthly data transfer rates for the instance.

iBundleId :: Lens' Instance (Maybe Text)

The bundle for the instance (e.g., micro_1_0 ).

iName :: Lens' Instance (Maybe Text)

The name the user gave the instance (e.g., Amazon_Linux-1GB-Virginia-1 ).

iSupportCode :: Lens' Instance (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

iBlueprintId :: Lens' Instance (Maybe Text)

The blueprint ID (e.g., os_amlinux_2016_03 ).

iPrivateIPAddress :: Lens' Instance (Maybe Text)

The private IP address of the instance.

iBlueprintName :: Lens' Instance (Maybe Text)

The friendly name of the blueprint (e.g., Amazon Linux ).

iIsStaticIP :: Lens' Instance (Maybe Bool)

A Boolean value indicating whether this instance has a static IP assigned to it.

iPublicIPAddress :: Lens' Instance (Maybe Text)

The public IP address of the instance.

iHardware :: Lens' Instance (Maybe InstanceHardware)

The size of the vCPU and the amount of RAM for the instance.

InstanceAccessDetails

data InstanceAccessDetails

The parameters for gaining temporary access to one of your Amazon Lightsail instances.

See: instanceAccessDetails smart constructor.

instanceAccessDetails :: InstanceAccessDetails

Creates a value of InstanceAccessDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iadCertKey - For SSH access, the public key to use when accessing your instance For OpenSSH clients (e.g., command line SSH), you should save this value to tempkey-cert.pub .

iadIpAddress - The public IP address of the Amazon Lightsail instance.

iadPrivateKey - For SSH access, the temporary private key. For OpenSSH clients (e.g., command line SSH), you should save this value to tempkey ).

iadExpiresAt - For SSH access, the date on which the temporary keys expire.

iadUsername - The user name to use when logging in to the Amazon Lightsail instance.

iadProtocol - The protocol for these Amazon Lightsail instance access details.

iadPassword - For RDP access, the temporary password of the Amazon EC2 instance.

iadInstanceName - The name of this Amazon Lightsail instance.

iadCertKey :: Lens' InstanceAccessDetails (Maybe Text)

For SSH access, the public key to use when accessing your instance For OpenSSH clients (e.g., command line SSH), you should save this value to tempkey-cert.pub .

iadIpAddress :: Lens' InstanceAccessDetails (Maybe Text)

The public IP address of the Amazon Lightsail instance.

iadPrivateKey :: Lens' InstanceAccessDetails (Maybe Text)

For SSH access, the temporary private key. For OpenSSH clients (e.g., command line SSH), you should save this value to tempkey ).

iadExpiresAt :: Lens' InstanceAccessDetails (Maybe UTCTime)

For SSH access, the date on which the temporary keys expire.

iadUsername :: Lens' InstanceAccessDetails (Maybe Text)

The user name to use when logging in to the Amazon Lightsail instance.

iadProtocol :: Lens' InstanceAccessDetails (Maybe InstanceAccessProtocol)

The protocol for these Amazon Lightsail instance access details.

iadPassword :: Lens' InstanceAccessDetails (Maybe Text)

For RDP access, the temporary password of the Amazon EC2 instance.

iadInstanceName :: Lens' InstanceAccessDetails (Maybe Text)

The name of this Amazon Lightsail instance.

InstanceHardware

data InstanceHardware

Describes the hardware for the instance.

See: instanceHardware smart constructor.

instanceHardware :: InstanceHardware

Creates a value of InstanceHardware with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ihCpuCount - The number of vCPUs the instance has.

ihDisks - The disks attached to the instance.

ihRamSizeInGb - The amount of RAM in GB on the instance (e.g., 1.0 ).

ihCpuCount :: Lens' InstanceHardware (Maybe Int)

The number of vCPUs the instance has.

ihDisks :: Lens' InstanceHardware [Disk]

The disks attached to the instance.

ihRamSizeInGb :: Lens' InstanceHardware (Maybe Double)

The amount of RAM in GB on the instance (e.g., 1.0 ).

InstanceNetworking

data InstanceNetworking

Describes monthly data transfer rates and port information for an instance.

See: instanceNetworking smart constructor.

instanceNetworking :: InstanceNetworking

Creates a value of InstanceNetworking with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

inMonthlyTransfer - The amount of data in GB allocated for monthly data transfers.

inPorts - An array of key-value pairs containing information about the ports on the instance.

inMonthlyTransfer :: Lens' InstanceNetworking (Maybe MonthlyTransfer)

The amount of data in GB allocated for monthly data transfers.

inPorts :: Lens' InstanceNetworking [InstancePortInfo]

An array of key-value pairs containing information about the ports on the instance.

InstancePortInfo

data InstancePortInfo

Describes information about the instance ports.

See: instancePortInfo smart constructor.

instancePortInfo :: InstancePortInfo

Creates a value of InstancePortInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ipiFromPort - The first port in the range.

ipiCommonName - The common name.

ipiProtocol - The protocol being used. Can be one of the following. * tcp - Transmission Control Protocol (TCP) provides reliable, ordered, and error-checked delivery of streamed data between applications running on hosts communicating by an IP network. If you have an application that doesn't require reliable data stream service, use UDP instead. * all - All transport layer protocol types. For more general information, see Transport layer on Wikipedia. * udp - With User Datagram Protocol (UDP), computer applications can send messages (or datagrams) to other hosts on an Internet Protocol (IP) network. Prior communications are not required to set up transmission channels or data paths. Applications that don't require reliable data stream service can use UDP, which provides a connectionless datagram service that emphasizes reduced latency over reliability. If you do require reliable data stream service, use TCP instead.

ipiAccessDirection - The access direction (inbound or outbound ).

ipiAccessType - The type of access (Public or Private ).

ipiToPort - The last port in the range.

ipiAccessFrom - The location from which access is allowed (e.g., Anywhere (0.0.0.0/0) ).

ipiFromPort :: Lens' InstancePortInfo (Maybe Natural)

The first port in the range.

ipiCommonName :: Lens' InstancePortInfo (Maybe Text)

The common name.

ipiProtocol :: Lens' InstancePortInfo (Maybe NetworkProtocol)

The protocol being used. Can be one of the following. * tcp - Transmission Control Protocol (TCP) provides reliable, ordered, and error-checked delivery of streamed data between applications running on hosts communicating by an IP network. If you have an application that doesn't require reliable data stream service, use UDP instead. * all - All transport layer protocol types. For more general information, see Transport layer on Wikipedia. * udp - With User Datagram Protocol (UDP), computer applications can send messages (or datagrams) to other hosts on an Internet Protocol (IP) network. Prior communications are not required to set up transmission channels or data paths. Applications that don't require reliable data stream service can use UDP, which provides a connectionless datagram service that emphasizes reduced latency over reliability. If you do require reliable data stream service, use TCP instead.

ipiAccessDirection :: Lens' InstancePortInfo (Maybe AccessDirection)

The access direction (inbound or outbound ).

ipiAccessType :: Lens' InstancePortInfo (Maybe PortAccessType)

The type of access (Public or Private ).

ipiToPort :: Lens' InstancePortInfo (Maybe Natural)

The last port in the range.

ipiAccessFrom :: Lens' InstancePortInfo (Maybe Text)

The location from which access is allowed (e.g., Anywhere (0.0.0.0/0) ).

InstancePortState

data InstancePortState

Describes the port state.

See: instancePortState smart constructor.

instancePortState :: InstancePortState

Creates a value of InstancePortState with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ipsFromPort - The first port in the range.

ipsState - Specifies whether the instance port is open or closed .

ipsProtocol - The protocol being used. Can be one of the following. * tcp - Transmission Control Protocol (TCP) provides reliable, ordered, and error-checked delivery of streamed data between applications running on hosts communicating by an IP network. If you have an application that doesn't require reliable data stream service, use UDP instead. * all - All transport layer protocol types. For more general information, see Transport layer on Wikipedia. * udp - With User Datagram Protocol (UDP), computer applications can send messages (or datagrams) to other hosts on an Internet Protocol (IP) network. Prior communications are not required to set up transmission channels or data paths. Applications that don't require reliable data stream service can use UDP, which provides a connectionless datagram service that emphasizes reduced latency over reliability. If you do require reliable data stream service, use TCP instead.

ipsToPort - The last port in the range.

ipsFromPort :: Lens' InstancePortState (Maybe Natural)

The first port in the range.

ipsState :: Lens' InstancePortState (Maybe PortState)

Specifies whether the instance port is open or closed .

ipsProtocol :: Lens' InstancePortState (Maybe NetworkProtocol)

The protocol being used. Can be one of the following. * tcp - Transmission Control Protocol (TCP) provides reliable, ordered, and error-checked delivery of streamed data between applications running on hosts communicating by an IP network. If you have an application that doesn't require reliable data stream service, use UDP instead. * all - All transport layer protocol types. For more general information, see Transport layer on Wikipedia. * udp - With User Datagram Protocol (UDP), computer applications can send messages (or datagrams) to other hosts on an Internet Protocol (IP) network. Prior communications are not required to set up transmission channels or data paths. Applications that don't require reliable data stream service can use UDP, which provides a connectionless datagram service that emphasizes reduced latency over reliability. If you do require reliable data stream service, use TCP instead.

ipsToPort :: Lens' InstancePortState (Maybe Natural)

The last port in the range.

InstanceSnapshot

data InstanceSnapshot

Describes the snapshot of the virtual private server, or instance .

See: instanceSnapshot smart constructor.

instanceSnapshot :: InstanceSnapshot

Creates a value of InstanceSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

insFromBlueprintId - The blueprint ID from which you created the snapshot (e.g., os_debian_8_3 ). A blueprint is a virtual private server (or instance ) image used to create instances quickly.

insState - The state the snapshot is in.

insResourceType - The type of resource (usually InstanceSnapshot ).

insArn - The Amazon Resource Name (ARN) of the snapshot (e.g., arn:aws:lightsail:us-east-1:123456789101:InstanceSnapshot/d23b5706-3322-4d83-81e5-12345EXAMPLE ).

insCreatedAt - The timestamp when the snapshot was created (e.g., 1479907467.024 ).

insLocation - The region name and availability zone where you created the snapshot.

insProgress - The progress of the snapshot.

insName - The name of the snapshot.

insFromBundleId - The bundle ID from which you created the snapshot (e.g., micro_1_0 ).

insSizeInGb - The size in GB of the SSD.

insSupportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

insFromInstanceARN - The Amazon Resource Name (ARN) of the instance from which the snapshot was created (e.g., arn:aws:lightsail:us-east-1:123456789101:Instance/64b8404c-ccb1-430b-8daf-12345EXAMPLE ).

insFromInstanceName - The instance from which the snapshot was created.

insFromBlueprintId :: Lens' InstanceSnapshot (Maybe Text)

The blueprint ID from which you created the snapshot (e.g., os_debian_8_3 ). A blueprint is a virtual private server (or instance ) image used to create instances quickly.

insState :: Lens' InstanceSnapshot (Maybe InstanceSnapshotState)

The state the snapshot is in.

insResourceType :: Lens' InstanceSnapshot (Maybe ResourceType)

The type of resource (usually InstanceSnapshot ).

insArn :: Lens' InstanceSnapshot (Maybe Text)

The Amazon Resource Name (ARN) of the snapshot (e.g., arn:aws:lightsail:us-east-1:123456789101:InstanceSnapshot/d23b5706-3322-4d83-81e5-12345EXAMPLE ).

insCreatedAt :: Lens' InstanceSnapshot (Maybe UTCTime)

The timestamp when the snapshot was created (e.g., 1479907467.024 ).

insLocation :: Lens' InstanceSnapshot (Maybe ResourceLocation)

The region name and availability zone where you created the snapshot.

insProgress :: Lens' InstanceSnapshot (Maybe Text)

The progress of the snapshot.

insName :: Lens' InstanceSnapshot (Maybe Text)

The name of the snapshot.

insFromBundleId :: Lens' InstanceSnapshot (Maybe Text)

The bundle ID from which you created the snapshot (e.g., micro_1_0 ).

insSizeInGb :: Lens' InstanceSnapshot (Maybe Int)

The size in GB of the SSD.

insSupportCode :: Lens' InstanceSnapshot (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

insFromInstanceARN :: Lens' InstanceSnapshot (Maybe Text)

The Amazon Resource Name (ARN) of the instance from which the snapshot was created (e.g., arn:aws:lightsail:us-east-1:123456789101:Instance/64b8404c-ccb1-430b-8daf-12345EXAMPLE ).

insFromInstanceName :: Lens' InstanceSnapshot (Maybe Text)

The instance from which the snapshot was created.

InstanceState

data InstanceState

Describes the virtual private server (or instance ) status.

See: instanceState smart constructor.

instanceState :: InstanceState

Creates a value of InstanceState with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

isName - The state of the instance (e.g., running or pending ).

isCode - The status code for the instance.

isName :: Lens' InstanceState (Maybe Text)

The state of the instance (e.g., running or pending ).

isCode :: Lens' InstanceState (Maybe Int)

The status code for the instance.

KeyPair

data KeyPair

Describes the SSH key pair.

See: keyPair smart constructor.

keyPair :: KeyPair

Creates a value of KeyPair with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kpResourceType - The resource type (usually KeyPair ).

kpArn - The Amazon Resource Name (ARN) of the key pair (e.g., arn:aws:lightsail:us-east-1:123456789101:KeyPair/05859e3d-331d-48ba-9034-12345EXAMPLE ).

kpCreatedAt - The timestamp when the key pair was created (e.g., 1479816991.349 ).

kpLocation - The region name and Availability Zone where the key pair was created.

kpFingerprint - The RSA fingerprint of the key pair.

kpName - The friendly name of the SSH key pair.

kpSupportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

kpResourceType :: Lens' KeyPair (Maybe ResourceType)

The resource type (usually KeyPair ).

kpArn :: Lens' KeyPair (Maybe Text)

The Amazon Resource Name (ARN) of the key pair (e.g., arn:aws:lightsail:us-east-1:123456789101:KeyPair/05859e3d-331d-48ba-9034-12345EXAMPLE ).

kpCreatedAt :: Lens' KeyPair (Maybe UTCTime)

The timestamp when the key pair was created (e.g., 1479816991.349 ).

kpLocation :: Lens' KeyPair (Maybe ResourceLocation)

The region name and Availability Zone where the key pair was created.

kpFingerprint :: Lens' KeyPair (Maybe Text)

The RSA fingerprint of the key pair.

kpName :: Lens' KeyPair (Maybe Text)

The friendly name of the SSH key pair.

kpSupportCode :: Lens' KeyPair (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

MetricDatapoint

data MetricDatapoint

Describes the metric data point.

See: metricDatapoint smart constructor.

metricDatapoint :: MetricDatapoint

Creates a value of MetricDatapoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdSampleCount - The sample count.

mdMaximum - The maximum.

mdAverage - The average.

mdMinimum - The minimum.

mdSum - The sum.

mdTimestamp - The timestamp (e.g., 1479816991.349 ).

mdUnit - The unit.

mdSampleCount :: Lens' MetricDatapoint (Maybe Double)

The sample count.

mdMaximum :: Lens' MetricDatapoint (Maybe Double)

The maximum.

mdAverage :: Lens' MetricDatapoint (Maybe Double)

The average.

mdMinimum :: Lens' MetricDatapoint (Maybe Double)

The minimum.

mdSum :: Lens' MetricDatapoint (Maybe Double)

The sum.

mdTimestamp :: Lens' MetricDatapoint (Maybe UTCTime)

The timestamp (e.g., 1479816991.349 ).

mdUnit :: Lens' MetricDatapoint (Maybe MetricUnit)

The unit.

MonthlyTransfer

data MonthlyTransfer

Describes the monthly data transfer in and out of your virtual private server (or instance ).

See: monthlyTransfer smart constructor.

monthlyTransfer :: MonthlyTransfer

Creates a value of MonthlyTransfer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mtGbPerMonthAllocated - The amount allocated per month (in GB).

mtGbPerMonthAllocated :: Lens' MonthlyTransfer (Maybe Int)

The amount allocated per month (in GB).

Operation

data Operation

Describes the API operation.

See: operation smart constructor.

operation :: Operation

Creates a value of Operation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

oStatus - The status of the operation.

oOperationDetails - Details about the operation (e.g., Debian-1GB-Virginia-1 ).

oResourceType - The resource type.

oCreatedAt - The timestamp when the operation was initialized (e.g., 1479816991.349 ).

oResourceName - The resource name.

oLocation - The region and Availability Zone.

oStatusChangedAt - The timestamp when the status was changed (e.g., 1479816991.349 ).

oErrorDetails - The error details.

oErrorCode - The error code.

oId - The ID of the operation.

oOperationType - The type of operation.

oIsTerminal - A Boolean value indicating whether the operation is terminal.

oStatus :: Lens' Operation (Maybe OperationStatus)

The status of the operation.

oOperationDetails :: Lens' Operation (Maybe Text)

Details about the operation (e.g., Debian-1GB-Virginia-1 ).

oResourceType :: Lens' Operation (Maybe ResourceType)

The resource type.

oCreatedAt :: Lens' Operation (Maybe UTCTime)

The timestamp when the operation was initialized (e.g., 1479816991.349 ).

oResourceName :: Lens' Operation (Maybe Text)

The resource name.

oLocation :: Lens' Operation (Maybe ResourceLocation)

The region and Availability Zone.

oStatusChangedAt :: Lens' Operation (Maybe UTCTime)

The timestamp when the status was changed (e.g., 1479816991.349 ).

oErrorDetails :: Lens' Operation (Maybe Text)

The error details.

oErrorCode :: Lens' Operation (Maybe Text)

The error code.

oId :: Lens' Operation (Maybe Text)

The ID of the operation.

oOperationType :: Lens' Operation (Maybe OperationType)

The type of operation.

oIsTerminal :: Lens' Operation (Maybe Bool)

A Boolean value indicating whether the operation is terminal.

PortInfo

data PortInfo

Describes information about the ports on your virtual private server (or instance ).

See: portInfo smart constructor.

portInfo :: PortInfo

Creates a value of PortInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

piFromPort - The first port in the range.

piProtocol - The protocol.

piToPort - The last port in the range.

piFromPort :: Lens' PortInfo (Maybe Natural)

The first port in the range.

piProtocol :: Lens' PortInfo (Maybe NetworkProtocol)

The protocol.

piToPort :: Lens' PortInfo (Maybe Natural)

The last port in the range.

RegionInfo

data RegionInfo

Describes the AWS Region.

See: regionInfo smart constructor.

regionInfo :: RegionInfo

Creates a value of RegionInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

riAvailabilityZones - The Availability Zones. Follows the format us-east-1a (case-sensitive).

riName - The region name (e.g., us-east-1 ).

riDisplayName - The display name (e.g., Virginia ).

riContinentCode - The continent code (e.g., NA , meaning North America).

riDescription - The description of the AWS Region (e.g., This region is recommended to serve users in the eastern United States and eastern Canada ).

riAvailabilityZones :: Lens' RegionInfo [AvailabilityZone]

The Availability Zones. Follows the format us-east-1a (case-sensitive).

riName :: Lens' RegionInfo (Maybe RegionName)

The region name (e.g., us-east-1 ).

riDisplayName :: Lens' RegionInfo (Maybe Text)

The display name (e.g., Virginia ).

riContinentCode :: Lens' RegionInfo (Maybe Text)

The continent code (e.g., NA , meaning North America).

riDescription :: Lens' RegionInfo (Maybe Text)

The description of the AWS Region (e.g., This region is recommended to serve users in the eastern United States and eastern Canada ).

ResourceLocation

data ResourceLocation

Describes the resource location.

See: resourceLocation smart constructor.

resourceLocation :: ResourceLocation

Creates a value of ResourceLocation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rlRegionName - The AWS Region name.

rlAvailabilityZone - The Availability Zone. Follows the format us-east-1a (case-sensitive).

rlRegionName :: Lens' ResourceLocation (Maybe RegionName)

The AWS Region name.

rlAvailabilityZone :: Lens' ResourceLocation (Maybe Text)

The Availability Zone. Follows the format us-east-1a (case-sensitive).

StaticIP

data StaticIP

Describes the static IP.

See: staticIP smart constructor.

staticIP :: StaticIP

Creates a value of StaticIP with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

siIpAddress - The static IP address.

siResourceType - The resource type (usually StaticIp ).

siArn - The Amazon Resource Name (ARN) of the static IP (e.g., arn:aws:lightsail:us-east-1:123456789101:StaticIp/9cbb4a9e-f8e3-4dfe-b57e-12345EXAMPLE ).

siCreatedAt - The timestamp when the static IP was created (e.g., 1479735304.222 ).

siLocation - The region and Availability Zone where the static IP was created.

siIsAttached - A Boolean value indicating whether the static IP is attached.

siName - The name of the static IP (e.g., StaticIP-Virginia-EXAMPLE ).

siSupportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

siAttachedTo - The instance where the static IP is attached (e.g., Amazon_Linux-1GB-Virginia-1 ).

siIpAddress :: Lens' StaticIP (Maybe Text)

The static IP address.

siResourceType :: Lens' StaticIP (Maybe ResourceType)

The resource type (usually StaticIp ).

siArn :: Lens' StaticIP (Maybe Text)

The Amazon Resource Name (ARN) of the static IP (e.g., arn:aws:lightsail:us-east-1:123456789101:StaticIp/9cbb4a9e-f8e3-4dfe-b57e-12345EXAMPLE ).

siCreatedAt :: Lens' StaticIP (Maybe UTCTime)

The timestamp when the static IP was created (e.g., 1479735304.222 ).

siLocation :: Lens' StaticIP (Maybe ResourceLocation)

The region and Availability Zone where the static IP was created.

siIsAttached :: Lens' StaticIP (Maybe Bool)

A Boolean value indicating whether the static IP is attached.

siName :: Lens' StaticIP (Maybe Text)

The name of the static IP (e.g., StaticIP-Virginia-EXAMPLE ).

siSupportCode :: Lens' StaticIP (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

siAttachedTo :: Lens' StaticIP (Maybe Text)

The instance where the static IP is attached (e.g., Amazon_Linux-1GB-Virginia-1 ).